git-health-analyser/backend/pyproject.toml
[project]
name = "github-health-analyzer"
version = "0.1.0"
description = "Analyze GitHub repository health and generate actionable insights"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "celery[redis]>=5.3.6",
    "redis>=5.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-jose[cryptography]>=3.3.0",
    "cryptography>=42.0.0",
    "slowapi>=0.1.9",
    "sentry-sdk[fastapi]>=1.40.0",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.2.0",
]

[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.ruff]
target-version = "py311"
line-length = 100
